refactor: reorganize folder structure for components
